+++ CHANGELOG.md
@@ -34,6 +34,8 @@ as 100% strength.
   - Option to enable or disable this.
 - A second set of buttons at the bottom of tips pages.
   - Option to enable or disable this.
+- Sub option to the Focus First Skills Option to set the priority order for which
+of the lists at the top of the tree is initially focused.
 
 ### Changed
 - Grammar skills are now separated in the crowns info breakdown. There are